remove useMessages in playground
