+++ packages/playground-ui/src/domains/agents/components/agent-chat.tsx
@@ -4,20 +4,48 @@ import { MastraRuntimeProvider } from '@/services/mastra-runtime-provider';
 import { ChatProps } from '@/types';
 import { useAgentSettings } from '../context/agent-context';
 import { usePlaygroundStore } from '@/store/playground-store';
+import { useAgentMessages } from '@/hooks/use-agent-messages';
+import { MastraUIMessage } from '@mastra/react';
+import { useEffect } from 'react';
 
 export const AgentChat = ({
   agentId,
   agentName,
   threadId,
-  initialMessages,
-  initialLegacyMessages,
   memory,
   refreshThreadList,
   modelVersion,
   modelList,
-}: ChatProps) => {
+  messageId,
+}: Omit<ChatProps, 'initialMessages' | 'initialLegacyMessages'> & { messageId?: string }) => {
   const { settings } = useAgentSettings();
   const { runtimeContext } = usePlaygroundStore();
+  const { data: messages, isLoading: isMessagesLoading } = useAgentMessages({
+    agentId: agentId,
+    threadId: threadId ?? '',
+    memory: memory ?? false,
+  });
+
+  // Handle scrolling to message after navigation
+  useEffect(() => {
+    if (messageId && messages && !isMessagesLoading) {
+      // Small delay to ensure DOM is ready
+      setTimeout(() => {
+        const messageElement = document.querySelector(`[data-message-id="${messageId}"]`);
+        if (messageElement) {
+          messageElement.scrollIntoView({ behavior: 'smooth', block: 'center' });
+          messageElement.classList.add('bg-surface4');
+          setTimeout(() => {
+            messageElement.classList.remove('bg-surface4');
+          }, 2000);
+        }
+      }, 100);
+    }
+  }, [messageId, messages, isMessagesLoading]);
+
+  if (isMessagesLoading) {
+    return null;
+  }
 
   return (
     <MastraRuntimeProvider
@@ -25,8 +53,8 @@ export const AgentChat = ({
       agentName={agentName}
       modelVersion={modelVersion}
       threadId={threadId}
-      initialMessages={initialMessages}
-      initialLegacyMessages={initialLegacyMessages}
+      initialMessages={(messages?.uiMessages || []) as MastraUIMessage[]}
+      initialLegacyMessages={messages?.legacyMessages || []}
       memory={memory}
       refreshThreadList={refreshThreadList}
       settings={settings}
