+++ packages/playground-ui/src/hooks/use-agent-messages.ts
@@ -1,13 +1,21 @@
 import { useMastraClient } from '@mastra/react';
 import { useQuery } from '@tanstack/react-query';
 
-export const useAgentMessages = ({ threadId, agentId }: { threadId: string; agentId: string }) => {
+export const useAgentMessages = ({
+  threadId,
+  agentId,
+  memory,
+}: {
+  threadId: string;
+  agentId: string;
+  memory: boolean;
+}) => {
   const client = useMastraClient();
 
   return useQuery({
     queryKey: ['memory', 'messages', threadId, agentId],
     queryFn: () => client.getThreadMessages(threadId, { agentId }),
-    enabled: Boolean(threadId),
+    enabled: memory && Boolean(threadId),
     staleTime: 0,
     gcTime: 0,
     retry: false,
