+++ packages/playground/src/pages/agents/agent/index.tsx
@@ -8,12 +8,11 @@ import {
 import { useEffect } from 'react';
 import { useNavigate, useParams, useSearchParams } from 'react-router';
 import { v4 as uuid } from '@lukeed/uuid';
-import { MastraUIMessage } from '@mastra/react';
 
 import { AgentInformation } from '@/domains/agents/agent-information';
 import { AgentSidebar } from '@/domains/agents/agent-sidebar';
 import { useAgent } from '@/hooks/use-agents';
-import { useMemory, useMessages, useThreads } from '@/hooks/use-memory';
+import { useMemory, useThreads } from '@/hooks/use-memory';
 
 function Agent() {
   const { agentId, threadId } = useParams();
@@ -21,11 +20,6 @@ function Agent() {
   const { data: agent, isLoading: isAgentLoading } = useAgent(agentId!);
   const { data: memory } = useMemory(agentId!);
   const navigate = useNavigate();
-  const { data: messages, isLoading: isMessagesLoading } = useMessages({
-    agentId: agentId!,
-    threadId: threadId!,
-    memory: !!memory?.result,
-  });
   const {
     data: threads,
     isLoading: isThreadsLoading,
@@ -40,23 +34,7 @@ function Agent() {
     }
   }, [memory?.result, threadId]);
 
-  // Handle scrolling to message after navigation
-  useEffect(() => {
-    const messageId = searchParams.get('messageId');
-    if (messageId && messages && !isMessagesLoading) {
-      // Small delay to ensure DOM is ready
-      setTimeout(() => {
-        const messageElement = document.querySelector(`[data-message-id="${messageId}"]`);
-        if (messageElement) {
-          messageElement.scrollIntoView({ behavior: 'smooth', block: 'center' });
-          messageElement.classList.add('bg-surface4');
-          setTimeout(() => {
-            messageElement.classList.remove('bg-surface4');
-          }, 2000);
-        }
-      }, 100);
-    }
-  }, [searchParams, messages, isMessagesLoading]);
+  const messageId = searchParams.get('messageId') ?? undefined;
 
   if (isAgentLoading) {
     return null;
@@ -79,19 +57,16 @@ function Agent() {
             )}
 
             <div className="grid overflow-y-auto relative bg-surface1 py-4">
-              {isMessagesLoading ? null : (
-                <AgentChat
-                  agentId={agentId!}
-                  agentName={agent?.name}
-                  modelVersion={agent?.modelVersion}
-                  threadId={threadId!}
-                  initialMessages={(messages?.uiMessages || []) as MastraUIMessage[]}
-                  initialLegacyMessages={messages?.legacyMessages || []}
-                  memory={memory?.result}
-                  refreshThreadList={refreshThreads}
-                  modelList={agent?.modelList}
-                />
-              )}
+              <AgentChat
+                agentId={agentId!}
+                agentName={agent?.name}
+                modelVersion={agent?.modelVersion}
+                threadId={threadId!}
+                memory={memory?.result}
+                refreshThreadList={refreshThreads}
+                modelList={agent?.modelList}
+                messageId={messageId}
+              />
             </div>
 
             <AgentInformation agentId={agentId!} />
